requirements.txt
plexapi
PyYAML
requests
tqdm
